prov/rxr: Remove incorrect assert in rxr_cq_handle_data()
The assertion was put in place to verify that the correct amount of data
was copied into the user buffers when processing data pkt completions.
When an incoming message is matched to a buffer that can not fully
contain the payload, we truncate the message and report the truncation
to the application. With the check in the assertion, rx_entry->total_len
reflects the size of the data sent by the peer, obtained from the
initial RTS and data_pkt->hdr.seg_size is obtained from the data pkt
header, neither of which are aware of the receive buffer size, causing
the assertion to fail in case of truncation.

Change-Id: I108a0c2d1da4bcbe7072388e9a1a900da670502b
Signed-off-by: Raghu Raja <craghun@amazon.com>