prov/efa: Remove READ_ONCE usage
We need to avoid the Kernel READ_ONCE implementation and, in
reviewing the single usage in sub-cq polling, we do not believe
we need any special memory handling of the flags field to read
the phase bit.

It is clear that we do not need full READ_ONCE semantics; it
is ok if flags is read multiple times from memory in handling
a CQE, given that we're explicitly reading it multiple times
in the control path.  Reading the control flow of how CQEs
are processed confirms this behavior is ok.

The next question is whether we need READ_ONCE for compiler
optimization semantics (ie, preventing the compiler from
caching the value of flags in a for loop).  The for loop
in this case is pretty big; it would be in RxR or higher, but
IPO could mean the compiler still sees the for loop and all
its contents as a for loop.  However, in the body of that
for loop is efa_io_cdesc_common, which both calls
cqe_is_pending() and contains a read memory barrier, which
is also a compiler barrier.  The loop can therefore be thought
of as:

for (...)
  /* lots of sub_cq parsing */
  is_pending = cqe_is_pending(cqe, sub_cq->phase);
  rmb();
  if (is_pending) {
     consume and return;
  }

The rmb() acting as a compiler barrier means that the compiler
can't assume that sub_cq->phase hasn't changed, so even if it
could see through all the sub cq parsing code, the code would
still be correct from a "need to read phase each time through
the loop" point of view.

Signed-off-by: Brian Barrett <bbarrett@amazon.com>
Signed-off-by: Raghu Raja <craghun@amazon.com>