prov/shm: save completion address properly
The previous implementation was not saving the source address
properly if requested, either improperly casting the fi_addr_t
to a pointer or not passing in the correct address.

Change the completion function to take in a fi_addr_t instead of
void * and make sure it's passing in the correct value.

Signed-off-by: aingerson <alexia.ingerson@intel.com>